Extract Security Master coverage helpers into a partial
Increment 2c of the endpoint-monolith decomposition. Moves the Security Master
coverage helper cluster out of WorkstationEndpoints.cs into a new sibling partial
WorkstationEndpoints.SecurityCoverage.cs.

Moved (10 whole private static methods, 5984-6223): BuildSecurityCoverage,
BuildResolvedSecurityReferences, BuildMissingSecurityReferences,
BuildPositionSecurityLookup, BuildTradingPositionPayload, BuildInlineSecurityReference,
BuildSecurityCoverageReference, BuildSecurityCoverageReason, HasAuthoritativeSecurityMatch,
NeedsSecurityReview.

Left in core: MapReconciliationStatus (a reconciliation-status helper used by the
accounting run card, not security coverage) and the SecurityCoveragePreviewLimit const
(class-level, referenced by the moved helpers across the partial). Because
WorkstationEndpoints is a single partial class this is a behavior-preserving relocation:
the core-resident briefing/accounting callers still reach BuildSecurityCoverage with no
signature/route/DI change. Core drops 7488 -> 7248 lines; ratchet baseline retightened.
